etch-a-sketch-server/requirements.txt
backgroundremover==0.2.8
click==8.1.7
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==3.0.2
flask_cors==4.0.1
ipdb==0.13.13
matplotlib==3.8.4
networkx==3.2.1
onnxruntime==1.17.3
# onnxruntime-gpu==1.17.1
opencv-python==4.10.0.82
opencv-python-headless==4.9.0.80
pandas==2.2.2
Pillow==9.5.0
pygame==2.5.2
pypotrace==0.3
rembg==2.0.56
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
tqdm==4.66.4
transitions==0.9.1
vtracer==0.6.11